painter added

+++ lib/no_sql/entities.py
@@ -1,12 +1,16 @@
 from typing import TypedDict
+
+
 class UserColor(TypedDict):
     user: int
     color: int
 
+
 class UserCoordinateX(TypedDict):
     user: int
-    x: float
+    x: int
+
 
 class UserCoordinateY(TypedDict):
     user: int
-    y: float
+    y: int
